Add file handler for logger
The `fuzza` module now has a logging level of `DEBUG`. An additional
file handler for the logger, with the same logging level is added to
output logs to a file.

This outputs log message with a level of `DEBUG` or higher to the log
files, while message with a level of `INFO` or higher to stdout.